+++ gazebo/rendering/LogicalCameraVisual.cc
@@ -46,52 +46,52 @@ LogicalCameraVisual::~LogicalCameraVisual()
 /////////////////////////////////////////////////
 void LogicalCameraVisual::Load(const msgs::LogicalCameraSensor &_msg)
 {
-  double nearWidth = (tan(_msg.horizontal_fov()*0.5) * _msg.near());
+  double nearWidth = (tan(_msg.horizontal_fov()*0.5) * _msg.near_clip());
   double nearHeight = nearWidth / _msg.aspect_ratio();
 
-  double farWidth = (tan(_msg.horizontal_fov()*0.5) * _msg.far());
+  double farWidth = (tan(_msg.horizontal_fov()*0.5) * _msg.far_clip());
   double farHeight = farWidth / _msg.aspect_ratio();
 
   DynamicLines *line = this->CreateDynamicLine(RENDERING_LINE_LIST);
 
   // Draw the near clipping plane as a white rectangle
-  line->AddPoint(math::Vector3(_msg.near(), nearWidth, nearHeight));
-  line->AddPoint(math::Vector3(_msg.near(), nearWidth, -nearHeight));
+  line->AddPoint(math::Vector3(_msg.near_clip(), nearWidth, nearHeight));
+  line->AddPoint(math::Vector3(_msg.near_clip(), nearWidth, -nearHeight));
 
-  line->AddPoint(math::Vector3(_msg.near(), nearWidth, -nearHeight));
-  line->AddPoint(math::Vector3(_msg.near(), -nearWidth, -nearHeight));
+  line->AddPoint(math::Vector3(_msg.near_clip(), nearWidth, -nearHeight));
+  line->AddPoint(math::Vector3(_msg.near_clip(), -nearWidth, -nearHeight));
 
-  line->AddPoint(math::Vector3(_msg.near(), -nearWidth, -nearHeight));
-  line->AddPoint(math::Vector3(_msg.near(), -nearWidth, nearHeight));
+  line->AddPoint(math::Vector3(_msg.near_clip(), -nearWidth, -nearHeight));
+  line->AddPoint(math::Vector3(_msg.near_clip(), -nearWidth, nearHeight));
 
-  line->AddPoint(math::Vector3(_msg.near(), -nearWidth, nearHeight));
-  line->AddPoint(math::Vector3(_msg.near(), nearWidth, nearHeight));
+  line->AddPoint(math::Vector3(_msg.near_clip(), -nearWidth, nearHeight));
+  line->AddPoint(math::Vector3(_msg.near_clip(), nearWidth, nearHeight));
 
   // Draw the far clipping plane as a white rectangle
-  line->AddPoint(math::Vector3(_msg.far(), farWidth, farHeight));
-  line->AddPoint(math::Vector3(_msg.far(), farWidth, -farHeight));
+  line->AddPoint(math::Vector3(_msg.far_clip(), farWidth, farHeight));
+  line->AddPoint(math::Vector3(_msg.far_clip(), farWidth, -farHeight));
 
-  line->AddPoint(math::Vector3(_msg.far(), farWidth, -farHeight));
-  line->AddPoint(math::Vector3(_msg.far(), -farWidth, -farHeight));
+  line->AddPoint(math::Vector3(_msg.far_clip(), farWidth, -farHeight));
+  line->AddPoint(math::Vector3(_msg.far_clip(), -farWidth, -farHeight));
 
-  line->AddPoint(math::Vector3(_msg.far(), -farWidth, -farHeight));
-  line->AddPoint(math::Vector3(_msg.far(), -farWidth, farHeight));
+  line->AddPoint(math::Vector3(_msg.far_clip(), -farWidth, -farHeight));
+  line->AddPoint(math::Vector3(_msg.far_clip(), -farWidth, farHeight));
 
-  line->AddPoint(math::Vector3(_msg.far(), -farWidth, farHeight));
-  line->AddPoint(math::Vector3(_msg.far(), farWidth, farHeight));
+  line->AddPoint(math::Vector3(_msg.far_clip(), -farWidth, farHeight));
+  line->AddPoint(math::Vector3(_msg.far_clip(), farWidth, farHeight));
 
   // Connect the near and far clipping planes with lines
-  line->AddPoint(math::Vector3(_msg.near(), nearWidth, nearHeight));
-  line->AddPoint(math::Vector3(_msg.far(), farWidth, farHeight));
+  line->AddPoint(math::Vector3(_msg.near_clip(), nearWidth, nearHeight));
+  line->AddPoint(math::Vector3(_msg.far_clip(), farWidth, farHeight));
 
-  line->AddPoint(math::Vector3(_msg.near(), -nearWidth, nearHeight));
-  line->AddPoint(math::Vector3(_msg.far(), -farWidth, farHeight));
+  line->AddPoint(math::Vector3(_msg.near_clip(), -nearWidth, nearHeight));
+  line->AddPoint(math::Vector3(_msg.far_clip(), -farWidth, farHeight));
 
-  line->AddPoint(math::Vector3(_msg.near(), -nearWidth, -nearHeight));
-  line->AddPoint(math::Vector3(_msg.far(), -farWidth, -farHeight));
+  line->AddPoint(math::Vector3(_msg.near_clip(), -nearWidth, -nearHeight));
+  line->AddPoint(math::Vector3(_msg.far_clip(), -farWidth, -farHeight));
 
-  line->AddPoint(math::Vector3(_msg.near(), nearWidth, -nearHeight));
-  line->AddPoint(math::Vector3(_msg.far(), farWidth, -farHeight));
+  line->AddPoint(math::Vector3(_msg.near_clip(), nearWidth, -nearHeight));
+  line->AddPoint(math::Vector3(_msg.far_clip(), farWidth, -farHeight));
 
   line->setMaterial("Gazebo/WhiteGlow");
   line->setVisibilityFlags(GZ_VISIBILITY_GUI);
@@ -99,16 +99,17 @@ void LogicalCameraVisual::Load(const msgs::LogicalCameraSensor &_msg)
   // Draw green lines from the near clipping plane to the origin
   DynamicLines *sourceLine = this->CreateDynamicLine(RENDERING_LINE_LIST);
   sourceLine->AddPoint(math::Vector3::Zero);
-  sourceLine->AddPoint(math::Vector3(_msg.near(), nearWidth, nearHeight));
+  sourceLine->AddPoint(math::Vector3(_msg.near_clip(), nearWidth, nearHeight));
 
   sourceLine->AddPoint(math::Vector3::Zero);
-  sourceLine->AddPoint(math::Vector3(_msg.near(), -nearWidth, nearHeight));
+  sourceLine->AddPoint(math::Vector3(_msg.near_clip(), -nearWidth, nearHeight));
 
   sourceLine->AddPoint(math::Vector3::Zero);
-  sourceLine->AddPoint(math::Vector3(_msg.near(), -nearWidth, -nearHeight));
+  sourceLine->AddPoint(math::Vector3(
+        _msg.near_clip(), -nearWidth, -nearHeight));
 
   sourceLine->AddPoint(math::Vector3::Zero);
-  sourceLine->AddPoint(math::Vector3(_msg.near(), nearWidth, -nearHeight));
+  sourceLine->AddPoint(math::Vector3(_msg.near_clip(), nearWidth, -nearHeight));
 
   sourceLine->setMaterial("Gazebo/PurpleGlow");
   sourceLine->setVisibilityFlags(GZ_VISIBILITY_GUI);